+++ src/model/constantly/apiConstants.js
@@ -1,78 +1,3 @@
-export const VersionState = async () => {
-  const versionApi = "https://ddragon.leagueoflegends.com/api/versions.json";
-  
-  const response = await fetch(versionApi);
-  const resJson = await response.json();
-  
-  return resJson[0];
-};
-
-export async function ChampionApi() {
-  const ver = await VersionState();
-  const verUrl = "https://ddragon.leagueoflegends.com/cdn/" + ver;
-
-  const championDataUrl = "https://ddragon.leagueoflegends.com/cdn/" + ver + "/data/ko_KR/champion.json";
-  const response = await fetch(championDataUrl);
-  const resJson = await response.json();
-
-  let imgList = [];
-  let nameList = [];
-  let idList = [];
-  
-  Object.values(resJson.data).sort((a, b) => (a.name < b.name ? -1 : a.name > b.name ? 1 : 0))
-  .map((index) => {
-    imgList.push(verUrl + "/img/champion/" + index.image.full);
-    nameList.push(index.name);
-    idList.push(index.id);
-  });
-
-  let dataList = [];
-  dataList.push(imgList);
-  dataList.push(nameList);
-  dataList.push(idList);
-
-  return dataList;
-}
-
-export async function RuneApi() {
-  const ver = await VersionState();
-
-  const runeUrl = "https://ddragon.leagueoflegends.com/cdn/" + ver + "/data/ko_KR/runesReforged.json";
-  const response = await fetch(runeUrl);
-  const resJson = await response.json();
-
-  return resJson;
-}
-
-export async function ChampionLineStats() {
-  const dataUrl = "http://localhost:8100/park/getData";
-
-  const response = await fetch(dataUrl);
-  const resJson = await response.json();
-
-  return resJson.champData.data;
-}
-
-export async function ChampionDetailDataApi() {
-  const dataUrl = "http://localhost:8100/park/getData2";
-  // const dataUrl = "http://localhost:8100/park/" + champId;
-
-  const response = await fetch(dataUrl);
-  const resJson = await response.json();
-  
-  return resJson[0];
-}
-
-export async function ChampionDetailApi(champId){
-  const dataUrl = "http://localhost:8100/park/garen";
-  // const dataUrl = "http://localhost:8100/park/" + champId;
-
-  const response = await fetch(dataUrl);
-  const resJson = await response.json();
-  
-  return resJson.champData.data;
-}
-
 export const apiKey = "RGAPI-d7f2268a-7c6a-4551-b4bd-092cb9d35f94";
 export const apiKey1 = "RGAPI-d74a8161-2e24-4f5d-9d13-58fff80dac26";
 // NicknameUrl + encodedName[i] + "?api_key=" + apiKey (id, acountid puuid name profileIconId level)
